requirements.txt
requests
click
psycopg2-binary